feat(auth): replace role-based with permission-based middleware
- Add new ModulePermission middleware for granular permission checks
- Update routes to use permission middleware instead of role-based checks
- Add permission checks to admin panel navigation and settings menu
- Remove forum-related routes and links from main layout
- Restructure admin routes with specific permission requirements

+++ app/Modules/Admin/Http/routes.php
@@ -3,29 +3,29 @@
 use Illuminate\Support\Facades\Route;
 
 
-Route::middleware(['auth', 'role:Admin', 'web'])
+Route::middleware(['auth', 'permission:access.admin.panel', 'web'])
     ->prefix(strtolower('acp'))
     ->group(function () {
-        Route::get('/', [\Modules\Admin\Http\Controllers\AdminController::class, 'index'])->name('admin.index');
-        Route::get('/permissions', [\Modules\Admin\Http\Controllers\PermissionController::class, 'index'])->name('admin.permissions');
+        Route::get('/', [\Modules\Admin\Http\Controllers\AdminController::class, 'index'])->middleware('permission:access.admin.panel')->name('admin.index');
+        Route::get('/permissions', [\Modules\Admin\Http\Controllers\PermissionController::class, 'index'])->middleware('permission:manage.permissions')->name('admin.permissions');
         
         // Roles
-        Route::get('/roles', [\Modules\Admin\Http\Controllers\RoleController::class, 'index'])->name('admin.roles.index');
-        Route::get('/roles/create', [\Modules\Admin\Http\Controllers\RoleController::class, 'create'])->name('admin.roles.create');
-        Route::post('/roles', [\Modules\Admin\Http\Controllers\RoleController::class, 'store'])->name('admin.roles.store');
-        Route::get('/roles/{role}/edit', [\Modules\Admin\Http\Controllers\RoleController::class, 'edit'])->name('admin.roles.edit');
-        Route::put('/roles/{role}', [\Modules\Admin\Http\Controllers\RoleController::class, 'update'])->name('admin.roles.update');
-        Route::delete('/roles/{role}', [\Modules\Admin\Http\Controllers\RoleController::class, 'destroy'])->name('admin.roles.destroy');
+        Route::get('/roles', [\Modules\Admin\Http\Controllers\RoleController::class, 'index'])->middleware('permission:manage.roles')->name('admin.roles.index');
+        Route::get('/roles/create', [\Modules\Admin\Http\Controllers\RoleController::class, 'create'])->middleware('permission:manage.roles')->name('admin.roles.create');
+        Route::post('/roles', [\Modules\Admin\Http\Controllers\RoleController::class, 'store'])->middleware('permission:manage.roles')->name('admin.roles.store');
+        Route::get('/roles/{role}/edit', [\Modules\Admin\Http\Controllers\RoleController::class, 'edit'])->middleware('permission:manage.roles')->name('admin.roles.edit');
+        Route::put('/roles/{role}', [\Modules\Admin\Http\Controllers\RoleController::class, 'update'])->middleware('permission:manage.roles')->name('admin.roles.update');
+        Route::delete('/roles/{role}', [\Modules\Admin\Http\Controllers\RoleController::class, 'destroy'])->middleware('permission:manage.roles')->name('admin.roles.destroy');
         
         // Administrative Users
-        Route::get('/users', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'index'])->name('admin.users.index');
-        Route::get('/users/create', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'create'])->name('admin.users.create');
-        Route::post('/users', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'store'])->name('admin.users.store');
-        Route::get('/users/{user}/edit', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'edit'])->name('admin.users.edit');
-        Route::put('/users/{user}', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'update'])->name('admin.users.update');
-        Route::delete('/users/{user}', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'destroy'])->name('admin.users.destroy');
+        Route::get('/users', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'index'])->middleware('permission:manage.admin.users')->name('admin.users.index');
+        Route::get('/users/create', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'create'])->middleware('permission:manage.admin.users')->name('admin.users.create');
+        Route::post('/users', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'store'])->middleware('permission:manage.admin.users')->name('admin.users.store');
+        Route::get('/users/{user}/edit', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'edit'])->middleware('permission:manage.admin.users')->name('admin.users.edit');
+        Route::put('/users/{user}', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'update'])->middleware('permission:manage.admin.users')->name('admin.users.update');
+        Route::delete('/users/{user}', [\Modules\Admin\Http\Controllers\AdminUserController::class, 'destroy'])->middleware('permission:manage.admin.users')->name('admin.users.destroy');
     
         // Settings
-        Route::get('/settings', [\Modules\Admin\Http\Controllers\AdminSettingsController::class, 'index'])->name('admin.settings.index');
-        Route::post('/settings', [\Modules\Admin\Http\Controllers\AdminSettingsController::class, 'store'])->name('admin.settings.store');
+        Route::get('/settings', [\Modules\Admin\Http\Controllers\AdminSettingsController::class, 'index'])->middleware('permission:manage.settings')->name('admin.settings.index');
+        Route::post('/settings', [\Modules\Admin\Http\Controllers\AdminSettingsController::class, 'store'])->middleware('permission:manage.settings')->name('admin.settings.store');
     });
